deploying the new changes

+++ src/components/ui/Dialog/DeleteAccount.vue
@@ -1,12 +1,13 @@
 <script setup lang="ts">
 import { inject, ref } from "vue";
-import { useRouter } from "vue-router";
+import { useRouter,useRoute } from "vue-router";
 import { useToast } from "vue-toast-notification";
 import { loader } from "../../..";
 import indexedDB from "../../../indexedDB"
 
 const origin:any=inject("origin")
 const router=useRouter()
+const route=useRoute()
 const userdata:any=inject("userdata")
 const isLoading=ref(false)
 const wait=ref("")
@@ -15,14 +16,17 @@ const dialog_close=()=>{
     const dialogElement=document.getElementById("delete-account") as HTMLDialogElement
     dialogElement.close()
 };
-
+const props=defineProps<{
+    data:any,
+    fetchDetails:any
+}>()
 async function clear(){
     try {
         dialog_close()
         loader.on()
         isLoading.value=true
         wait.value="cursor-progress bg-gray-400"
-        const url=!userdata.username?`${origin}/api/groups/${userdata.email}`:`${origin}/api/accounts/${userdata.email}`
+        const url=!route.query.type?`${origin}/api/groups/${userdata.email}`:`${origin}/api/accounts/${route.query.email}`
         const response=await fetch(url,{
             method:"DELETE",
             headers:{
@@ -37,24 +41,27 @@ async function clear(){
             })
             loader.off()
         }else{
-            const request=await indexedDB()
-            const db:any=await request
-            const transaction=db.transaction("All_files","readwrite")
-            const fileStore=transaction.objectStore("All_files")
-            const fileEmail=fileStore.index("email")
-            const fileEmailKey = fileEmail.getAllKeys([`${userdata.email}`]);
-            fileEmailKey.onsuccess=()=>{
-                console.log(fileEmailKey.result)
-                fileEmailKey.result.forEach((item:any)=>{
-                    const deleteFiles=fileStore.delete(item)
-                })
-                dialog_close()
-                localStorage.removeItem("userdata")
-                router.back()
+            if(route.query.type){
+                const request=await indexedDB()
+                const db:any=await request
+                const transaction=db.transaction("All_files","readwrite")
+                const fileStore=transaction.objectStore("All_files")
+                const fileEmail=fileStore.index("email")
+                const fileEmailKey = fileEmail.getAllKeys([`${userdata.email}`]);
+                fileEmailKey.onsuccess=()=>{
+                    console.log(fileEmailKey.result)
+                    fileEmailKey.result.forEach((item:any)=>{
+                        const deleteFiles=fileStore.delete(item)
+                    })
+                    dialog_close()
+                    localStorage.removeItem("userdata")
+                    router.back()
+                }
+                fileEmailKey.onerror=()=>{
+                    console.log("error",fileEmailKey.result)
+                }
             }
-            fileEmailKey.onerror=()=>{
-                console.log("error",fileEmailKey.result)
-            }
+            props.fetchDetails()
         }
 
     } catch (error:any) {
@@ -77,7 +84,7 @@ async function clear(){
             <i class="icon pi pi-times text-lg hover:text-[#F45858]"></i>
         </button>
         <div class="flex flex-col w-full">
-            <p class="text-black mb-5 text-center max-sm:text-xs">You are about to delete account</p>
+            <p class="text-black mb-5 text-center max-sm:text-xs">You are about to delete <span v-if="data.group_ownership&&!route.query.type">{{ data.group_ownership }}</span> <span v-else-if="data.groupname">{{ data.groupname }}</span> <span v-else-if="route.query.type">account</span></p>
             <p class="text-red-500 mb-5 text-center text-sm max-sm:text-xs">Once you delete your account all your data would be lost</p>
             <div class="flex gap-6 max-md:text-sm justify-between">
                 <button :disabled="isLoading"  @click="clear" class="text-white bg-red-600 rounded-[10px] h-[40px] w-[120px]">
